refactor(nexus-web): Chat UI module Phase 1 scaffolding
Colocate chat route parsing and thread page under the chat app module,
matching Account/Files structure while preserving existing URLs.
