+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/components/shell/workspace-layout.tsx
@@ -11,7 +11,7 @@ import { MobileBottomNav } from './mobile/mobile-bottom-nav';
 import { MobileMoreSheet } from './mobile/mobile-more-sheet';
 import { MobileTopBar } from './mobile/mobile-top-bar';
 import { ChatExportButton } from '@/components/chat/chat-export-button';
-import { parseChatRoute } from './chat-route';
+import { parseChatRoute } from '@/app/workspace/[workspaceId]/chat/lib/chat-route';
 import { parseFilesRoute } from '@/app/workspace/[workspaceId]/files/lib/files-route';
 import { FilesSection } from './sidebar/files-section';
 import { useIsMobile } from '@/hooks/use-is-mobile';
